+++ pyhottop/pyhottop.py
@@ -30,6 +30,7 @@ import logging
 import serial
 import sys
 import time
+from scipy.stats import linregress
 
 py2 = sys.version[0] == '2'
 
@@ -39,6 +40,7 @@ else:
     from queue import Queue
 
 from threading import Thread, Event
+from collections import deque
 
 
 class InvalidInput(Exception):
@@ -123,6 +125,9 @@ class ControlProcess(Thread):
     :returns: ControlProces instance
     """
 
+    MAX_BOUND_TEMP = 500
+    MIN_BOUND_TEMP = 100
+
     def __init__(self, conn, config, q, logger, callback=None):
         """Extend threads to support more control logic."""
         Thread.__init__(self)
@@ -160,6 +165,9 @@ class ControlProcess(Thread):
         config[12] = self._config.get('main_fan', 0)
         config[16] = self._config.get('solenoid', 0)
         config[17] = self._config.get('drum_motor', 0)
+        if self._config.get('heater', 0) > 0:
+            # Override the user here since the drum MUST be on for heat
+            config[17] = 1
         config[18] = self._config.get('cooling_motor', 0)
         config[35] = sum([b for b in config[:35]]) & 0xFF
         return bytes(config)
@@ -223,16 +231,16 @@ class ControlProcess(Thread):
         self._conn.flushOutput()
         buffer = self._conn.read(36)
         if len(buffer) != 36:
-            self._log('Buffer length did not match 36')
+            self._log.debug('Buffer length did not match 36')
             if self._conn.isOpen():
                 self._log.debug('Closing connection')
                 self._conn.close()
                 self._read_settings(retry=True)
 
         check = self._validate_checksum(buffer)
-        if not check and (retry and self._retry_count <= 3):
-            if self._retry_count == 3:
-                self._log.debug('Retry count reached on buffer check')
+        if not check and (retry and self._retry_count <= 10):
+            if self._retry_count == 10:
+                self._log.error('Retry count reached on buffer check')
                 self._read_settings(retry=False)
             else:
                 self._retry_count += 1
@@ -250,8 +258,31 @@ class ControlProcess(Thread):
         settings['drum_motor'] = hex2int(buffer[17])
         settings['cooling_motor'] = hex2int(buffer[18])
         settings['chaff_tray'] = hex2int(buffer[19])
+        settings['buffer'] = buffer
         self._retry_count = 0
         return settings
+
+    def _valid_config(self, settings):
+        """Scan through the returned settings to ensure they appear sane.
+
+        There are time when the returned buffer has the proper information, but
+        the reading is inaccurate. When this happens, temperatures will swing
+        or system values will be set to improper values.
+
+        :param settings: Configuration derived from the buffer
+        :type settings: dict
+        :returns: bool
+        """
+        if ((int(settings['environment_temp']) > self.MAX_BOUND_TEMP or
+                int(settings['environment_temp']) < self.MIN_BOUND_TEMP) or
+            (int(settings['bean_temp']) > self.MAX_BOUND_TEMP or
+                int(settings['bean_temp']) < self.MIN_BOUND_TEMP)):
+            return False
+        binary = ['drum_motor', 'chaff_tray', 'solenoid', 'cooling_motor']
+        for item in binary:
+            if int(settings.get(item)) not in [0, 1]:
+                return False
+        return True
 
     def _wake_up(self):
         """Wake the machine up to avoid race conditions.
@@ -291,6 +322,7 @@ class ControlProcess(Thread):
 
         while not self.exit.is_set():
             settings = self._read_settings()
+            settings['valid'] = self._valid_config(settings)
             self._cb(settings)
 
             if self.cooldown.is_set():
@@ -301,7 +333,8 @@ class ControlProcess(Thread):
                 self._config['cooling_motor'] = 1
                 self._config['main_fan'] = 10
 
-            self._send_config()
+            if settings['valid']:
+                self._send_config()
             time.sleep(self._config['interval'])
 
     def drop(self):
@@ -336,7 +369,7 @@ class Hottop:
     STOPBITS = 1
     TIMEOUT = 1
     LOG_LEVEL = logging.DEBUG
-    INTERVAL = 1
+    INTERVAL = 0.6
 
     def __init__(self):
         """Start of the hottop."""
@@ -347,6 +380,7 @@ class Hottop:
         self._roast_start = None
         self._roast_end = None
         self._config = dict()
+        self._window = deque(list(), 5)
         self._q = Queue()
         self._init_controls()
 
@@ -452,6 +486,8 @@ class Hottop:
         self._roast['events'] = list()
         self._roast['last'] = dict()
         self._roast['record'] = False
+        self._roast['charge'] = None
+        self._roast['turning_point'] = None
 
     def _callback(self, data):
         """Processor callback to clean-up stream data.
@@ -479,14 +515,68 @@ class Hottop:
             self._roast['duration'] = timedelta2period(ct - st)
 
         if self._roast['record']:
-            self._roast['events'].append(copy.deepcopy(output))
+            copied = copy.deepcopy(output)
+            self._derive_charge(copied['config'])
+            self._derive_turning_point(copied['config'])
+            self._roast['events'].append(copied)
             self._roast['last'] = local
 
         if self._user_callback:
             self._log.debug("Passing data back to client handler")
             output['roast'] = self._roast
             output['roasting'] = self._roasting
-            self._user_callback(output)
+            if local.get('valid', True):
+                self._user_callback(output)
+
+    def _derive_charge(self, config):
+        """Use a temperature window to identify the roast charge.
+
+        The charge will manifest as a sudden downward trend on the temperature.
+        Once found, we save it and avoid overwriting. The charge is needed in
+        order to derive the turning point.
+
+        :param config: Current snapshot of the configuration
+        :type config: dict
+        :returns: None
+        """
+        if self._roast.get('charge'):
+            return None
+        self._window.append(config)
+        time, temp = list(), list()
+        for x in list(self._window):
+            time.append(x['time'])
+            temp.append(x['bean_temp'])
+        slope, intercept, r_value, p_value, std_err = linregress(time, temp)
+        if slope < 0:
+            self._roast['charge'] = self._roast['last']
+            self.add_roast_event({'event': 'Charge'})
+            return config
+        return None
+
+    def _derive_turning_point(self, config):
+        """Use a temperature window to identify the roast turning point.
+
+        Turning point relies on the charge being set first. We use the rolling
+        5-point window to measure slope. If we show a positive trend after
+        the charge, then the temperature has begun to turn.
+
+        :param config: Current snapshot of the configuration
+        :type config: dict
+        :returns: None
+        """
+        if not self._roast.get('charge') or self._roast.get('turning_point'):
+            return None
+        self._window.append(config)
+        time, temp = list(), list()
+        for x in list(self._window):
+            time.append(x['time'])
+            temp.append(x['bean_temp'])
+        slope, intercept, r_value, p_value, std_err = linregress(time, temp)
+        if slope > 0:
+            self._roast['turning_point'] = self._roast['last']
+            self.add_roast_event({'event': 'Turning Point'})
+            return config
+        return None
 
     def start(self, func=None):
         """Start the roaster control process.